refactor(common, usuario): se mueve la interfaz genérica a una carpeta common ya que será implementada por varios servicios
